Add Haskell solutions for problems 34 and 36
Also moved the numericPalindrome function to the common module
(needed in problem 4 and 36).

A bug in numericPalindrome was also fixed. Values such as 1011
would result in false positive dues to the `n < base` check
being applied to n = 1 and base = 10 with top = 10.
This guard check has been replaced with a conceptually
equivalent check on the top value.